skip cleanup code
fix #431
